Sort log files to be processed
It is much easier to skim over logs and search for some problems
if you know that files are processed in some particular order.

Plain lexicographical sort is fine for all files that have YYYYMMDD
as the non-constant part of the name: it coincides with sort by date
that is the most natural thing.

Signed-off-by: Adrian Coveney <adrian.coveney@stfc.ac.uk>