Update setup.py flow to perform basic setup standard commands
Now, `setup.py` will make use of the build step to install the package. Enabling separate flows for the build and install phases if specified.

We are currently NOT supporting a few standard commands that are out of scope or NOT being asked, listed below:
- `register`, `bdist_rpm`, `bdist_wininst`, `upload`.

Also, We do NOT support extra commands with the setup.py except a few commands like `develop`, `egg_info`, and `bdist_egg`.