Rewrote lines to match suggested improvement
.fromkeys is a classmethod since ages. Use it on the class dict instead
of a throwaway instance {}. The and/or was abusing boolean operators to
implement a ternary expression; it's idiomatic to just use a ternary as
'RecordIdentity' if f == 'RecordId' else f. Constructing a list by
iteration only to iterate over the result is superfluous. Directly
iterate instead and remove the intermediate list.