Remove "caching" of keys in record.py
There is no advantage to "caching" the keys before the loop. The loop
only changes the values, not the keys, of the dictionary. Since
searching for a key in a dict ( O(1) hash lookup) is faster than in a
list ( O(n) search ) there is no reason to materialise the keys.