Replace optparse with argparse
Optparse is deprecated - no longer developed. Mostly a like-for-like
replacement, but we use vars on the argparse object to output the
values in a similar way to optparse.

The old 'version' option on the constructor is unavailable, and so it
has been replaced by adding an argument with the 'version' action.